keep the index in the input_data and save it into the config file
since it is so hard to keep index clean even with deepcopy to separate the input_data_queue and original_input_data. i change the design to keep the index in the input_data and save it into the config file. right now each input data will have a unique key.
so no need to manage the index removal and add-back. this pr will keep the index clean